feat : 추가 board password 기능

+++ intensive-immersion-training-board-project_cqrs_board_command/src/main/java/com/example/intensiveimmersiontrainingboardproject_cqrs_board_command/service/BoardServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.intensiveimmersiontrainingboardproject_cqrs_board_command.domain.Board;
 import com.example.intensiveimmersiontrainingboardproject_cqrs_board_command.dto.BoardDto;
+import com.example.intensiveimmersiontrainingboardproject_cqrs_board_command.dto.PasswordDto;
 import com.example.intensiveimmersiontrainingboardproject_cqrs_board_command.repository.BoardRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -22,8 +23,8 @@ public class BoardServiceImpl implements BoardService {
 		try {
 			Board boardEntity = Board.builder().title(boardDto.getTitle()).content(boardDto.getContent())
 				.createdDate(boardDto.getCreatedDate()).nickname(boardDto.getNickname()).modifiedDate(
-					boardDto.getModifiedDate())
-			.build();
+					boardDto.getModifiedDate()).password(boardDto.getPassword())
+				.build();
 			boardRepository.save(boardEntity);
 		} catch (Exception e) {
 			throw new Exception("저장에 실패했습니다.");
@@ -33,9 +34,15 @@ public class BoardServiceImpl implements BoardService {
 
 	@Override
 	@Transactional
-	public void deleteBoard(Long id) throws Exception {
+	public void deleteBoard(Long id, PasswordDto passwordDto) throws Exception {
 		try {
-			boardRepository.deleteBoardById(id);
+			Board board = boardRepository.findBoardById(id).orElse(null);
+			assert board != null;
+			if (passwordDto.getPassword().equals(board.getPassword())) {
+				boardRepository.deleteBoardById(id);
+			}else{
+				throw new IllegalArgumentException("비밀번호 이슈");
+			}
 		} catch (Exception e) {
 			throw new Exception("삭제에 실패했습니다.");
 		}
@@ -43,11 +50,11 @@ public class BoardServiceImpl implements BoardService {
 
 	@Override
 	@Transactional
-	public void updateTitleAndContent(Long id, String title,String content) throws Exception {
+	public void updateTitleAndContent(Long id, String title, String content) throws Exception {
 		Board board = boardRepository.findBoardById(id).orElse(null);
 		if (board != null) {
-			board.updateTitleAndContent(title,content);
-		}else {
+			board.updateTitleAndContent(title, content);
+		} else {
 			throw new Exception("이미 존재합니다.");
 		}
 	}
